When generating a unique ID for function overloads, ignore the return parameter. From the C++ standard, 1.311:
"the information about a function that participates in overload resolution (13.3): its parameter-type-list (8.3.5) and, if the function is a class member, the cv-qualifiers (if any) on the function itself and the class in which the member function is declared."

As functions can only be overloaded using the derived covariant approach, this works exactly as it did before.

Thanks to @david_meiners for raising this issue.